+++ flutter_sound/example/lib/demo/demo.dart
@@ -191,7 +191,7 @@ class _MyAppState extends State<Demo> {
     'https://flutter-sound.canardoux.xyz/extract/13.wav', // 'assets/samples/sample.wav',
     'https://flutter-sound.canardoux.xyz/extract/02.aiff', // 'assets/samples/sample.aiff',
     'https://flutter-sound.canardoux.xyz/extract/01-pcm.caf', // 'assets/samples/sample_pcm.caf',
-    'https://flutter-sound.canardoux.xyz/extract/extract/04.flac', // 'assets/samples/sample.flac',
+    'https://flutter-sound.canardoux.xyz/extract/04.flac', // 'assets/samples/sample.flac',
     'https://flutter-sound.canardoux.xyz/extract/06.mp4', // 'assets/samples/sample.mp4',
     'https://flutter-sound.canardoux.xyz/extract/03.amr', // 'assets/samples/sample.amr', // amrNB
     'https://flutter-sound.canardoux.xyz/extract/03.amr', // 'assets/samples/sample_xxx.amr', // amrWB