+++ flutter_sound/example/lib/livePlaybackWithoutBackPressure/live_playback_without_back_pressure.dart
@@ -65,13 +65,14 @@ class _LivePlaybackWithoutBackPressureState
     extends State<LivePlaybackWithoutBackPressure> {
   FlutterSoundPlayer? _mPlayer = FlutterSoundPlayer();
   bool _mPlayerIsInited = false;
+  bool _mEnableVoiceProcessing = false;
 
   @override
   void initState() {
     super.initState();
     // Be careful : openAudioSession return a Future.
     // Do not access your FlutterSoundPlayer or FlutterSoundRecorder before the completion of the Future
-    _mPlayer!.openPlayer().then((value) {
+    _mPlayer!.openPlayer(enableVoiceProcessing: _mEnableVoiceProcessing).then((value) {
       setState(() {
         _mPlayerIsInited = true;
       });
@@ -153,7 +154,7 @@ class _LivePlaybackWithoutBackPressureState
           Container(
             margin: const EdgeInsets.all(3),
             padding: const EdgeInsets.all(3),
-            height: 80,
+            height: 120,
             width: double.infinity,
             alignment: Alignment.center,
             decoration: BoxDecoration(
@@ -163,7 +164,7 @@ class _LivePlaybackWithoutBackPressureState
                 width: 3,
               ),
             ),
-            child: Row(children: [
+            child: Column(children: [Row(children: [
               ElevatedButton(
                 onPressed: getPlaybackFn(),
                 //color: Colors.white,
@@ -177,8 +178,35 @@ class _LivePlaybackWithoutBackPressureState
                   ? 'Playback in progress'
                   : 'Player is stopped'),
             ]),
-          ),
-        ],
+           Row(children: [
+             Checkbox(
+                checkColor: Colors.white,
+                //fillColor: Colors.white,
+                value: _mEnableVoiceProcessing,
+                onChanged: (bool? value) {
+                    _mPlayer!.closePlayer().then((v) {
+                      _mPlayerIsInited = false;
+                      _mEnableVoiceProcessing = value!;
+                      _mPlayer!.openPlayer(enableVoiceProcessing: _mEnableVoiceProcessing);}).then((value) {
+                        setState(() {
+                        _mPlayerIsInited = true;
+                      });
+                    }
+
+                    );
+      }
+                  ),
+             const Text("EnableVoiceProcessing")
+
+           ])
+    ]
+             ),
+
+
+
+         ),
+        ]
+
       );
     }