+++ flutter_sound/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 9.3.7
+## 9.3.8
 
 - ## The [CHANGELOG file is here](https://flutter-sound.canardoux.xyz/changelog.html)