+++ flutter_sound_platform_interface/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 9.3.7
+## 9.3.8
 
 - Please [look to this](https://flutter-sound.canardoux.xyz/changelog.html)